refactor: extract alpha select state init, api init, event init to index.ts
